// Angularrepo/ascendion-fsd-jan-2025/mywork/workshop-app/src/app/workshops/models/ISession.ts
type Level = 'Basic' | 'Intermediate' | 'Advanced';

export interface ISession {
    id: number;
    workshopId: number;
    sequenceId: number;
    name: string;
    speaker: string;
    duration: number;
    level: Level;
    abstract: string;
    upvoteCount: number;
}

export type {
    Level,
    ISession as default
};